# docs/api.yaml
swagger: '2.0'
info:
  title: 用户相关 API 接口
  version: 1.0.0
  description: 'ts -> utc+8 , Beijing Time <br/> salt -> "棘心夭夭，母氏劬劳。母氏圣善，我无令人。"'
host: 'localhost:8003'
basePath: /v1
schemes:
  - http
...
paths:
...
  /recommend/carer/hot:
    get:
      tags:
        - recommend
      summary: 推荐热门看护人
      operationId: recommend_carer_hot
      description: 推荐热门看护人
      produces:
        - application/json
      parameters:
        - in: query
          name: offset
          type: integer
          default: 0
          description: 偏移
        - in: query
          name: limit
          type: integer
          default: 5
          description: 数量
        - in: query
          name: city_id
          type: integer
          description: '城市编号'
        - in: query
          name: lat
          type: number
          format: 'decimal(10, 6)'
        - in: query
          name: lng
          type: number
          format: 'decimal(10, 6)'
        - in: query
          name: child_born
          type: integer
          description: '孩子是否已出生 1'
          enum:
            - 0 未出生
            - 1 已出生
        - in: query
          name: child_birth_ts
          type: integer
          description: '孩子出生日期 时间戳 北京时间 1523264525'
        - in: query
          name: child_gender
          type: integer
          enum:
            - 1 女
            - 2 男
        - in: query
          name: child_relation
          description: 用户和孩子的关系 2
          type: integer
          enum:
            - 0 未知
            - 1 妈妈
            - 2 爸爸
            - 3 爷爷
            - 4 奶奶
            - 5 外公
            - 6 外婆
            - 7 其他家人
        - in: query
          name: child_nickname
          type: string
          description: 孩子的昵称 卖马的秦琼
      responses:
        '200':
          description: returned status
          schema:
            type: object
            required:
              - status
              - data
            properties:
              status:
                $ref: '#/definitions/Status'
              data:
                type: array
                items:
                  $ref: '#/definitions/RecommendCarerResponseDataItem'
...
definitions:
  Status:
    type: object
    required:
      - code
      - msg
    properties:
      code:
        description: 错误码，成功为 0
        type: integer
        example: 0
      msg:
        description: 错误信息
        type: string
        example: ''
...
  UserVideo:
    type: object
    properties:
      key:
        type: string
        description: qiniu oss key
        example: user_id_123/avatar/2018/03/138/etag.jpg
      mime_type:
        type: string
        description: 文件类型 mimeType
        example: image/jpeg
      etag:
        type: string
        description: 使用 qiniu sdk 计算的 etag
        example: 7DsdkdFSKkdljfksQWET-TUI
      size:
        type: integer
        description: 视频文件大小(字节)
        example: 10240002
      width:
        type: integer
        description: 视频分辨率 宽
        example: 800
      height:
        type: integer
        description: 视频分辨率 高
        example: 600
      duration:
        type: number
        description: 视频播放时长
        example: 10.0
      cover_url:
        description: 视频快照截图 URL
        type: string
        format: url
        example: 'http://p5zmpa3g9.bkt.clouddn.com/cover.jpg'
      video_url:
        description: 视频URL
        type: string
        format: url
        example: 'http://p5zmpa3g9.bkt.clouddn.com/video.mp4'
  RecommendCarerResponseDataItem:
    type: object
    properties:
      user_id:
        type: string
        format: uuid
        description: 用户 id
        example: 123456
      nickname:
        type: string
        description: 昵称
        example: xxx的妈妈
      avatar_url:
        type: string
        format: url
        description: 头像图片 URL
        example: 'qiniu.com/image/2018/03/11/avatar.jpg'
      address:
        type: string
        description: 简短地址
        example:
      address_name:
        type: string
        description: user_address 里的 name 字段
        example:
      distance:
        type: string
        description: 距离
        example: 1.2km(821m)
      liked:
        type: boolean
        description: 是否赞过
        example: true
      like_count:
        type: integer
        description: 被点赞次数
        example: 100
      followed:
        type: boolean
        description: 是否收藏过
        example: false
      follow_count:
        type: integer
        description: 被收藏次数
        example: 28
      videos:
        type: array
        items:
          $ref: '#/definitions/UserVideo'
        description: 看护人的视频
...